+++ lib/logging_appenders.dart
@@ -1,5 +1,6 @@
 library logging_appenders;
 
+export 'src/base_appender.dart' show LoggingAppenders, BaseLogAppender;
 export 'src/logrecord_formatter.dart';
 export 'src/print_appender.dart';
 export 'src/remote/logzio_appender.dart';