+++ lib/src/remote/logzio_appender.dart
@@ -3,21 +3,34 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:dio/dio.dart';
-import 'package:logging/logging.dart';
+import 'package:logging_appenders/logging_appenders.dart';
+import 'package:logging_appenders/src/dummy_logger.dart';
 import 'package:logging_appenders/src/remote/base_remote_appender.dart';
 import 'package:meta/meta.dart';
 
 // ignore: unused_element
-final _logger = Logger('loki_logger');
+final _logger = DummyLogger('logging_appenders.logzio_appender');
 
+/// Appender which sends all logs to https://logz.io/
+/// Uses
 class LogzIoApiAppender extends BaseDioLogSender {
   LogzIoApiAppender({
+    LogRecordFormatter formatter,
     @required this.apiToken,
     @required this.labels,
-  });
+    this.url = 'https://listener.logz.io:8071/',
+    this.type = 'flutterlog',
+    int bufferSize,
+  })  : assert(apiToken != null),
+        assert(labels != null),
+        assert(url != null),
+        assert(type != null),
+        super(formatter: formatter, bufferSize: bufferSize);
 
+  final String url;
   final String apiToken;
   final Map<String, String> labels;
+  final String type;
 
   Dio _clientInstance;
 
@@ -26,6 +39,7 @@ class LogzIoApiAppender extends BaseDioLogSender {
   @override
   Future<void> sendLogEventsWithDio(List<LogEntry> entries,
       Map<String, String> userProperties, CancelToken cancelToken) {
+    _logger.finest('Sending logs to $url');
     final body = entries
         .map((entry) => {
               '@timestamp': entry.ts.toUtc().toIso8601String(),
@@ -38,7 +52,7 @@ class LogzIoApiAppender extends BaseDioLogSender {
         .join('\n');
     return _client
         .post<dynamic>(
-          'https://listener.logz.io:8071/?token=$apiToken&type=flutterlog',
+          '$url?token=$apiToken&type=$type',
           data: body,
           cancelToken: cancelToken,
           options: Options(